requirements.txt
flask==3.0.0
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.1.0
werkzeug==3.0.1
PyPDF2==3.0.1
